+++ core/ChangeLog.md
@@ -5,6 +5,8 @@ XP Framework ChangeLog
 
 ### Bugfixes
 
+* Fixed io.streams.TextReader class not reading first chunk correctly when
+  character set auto-detection fell back to iso-8859-1 (@thekid)
 * Fixed StreamTokenizer::pushBack() not working correctly if previously
   read all the way to the end - (@thekid)
 * Fixed issue #324 - Encoding not handled correctly in JsonDecoder; the